Add PR Update AsyncWebSocket.cpp #1142

+++ src/AsyncWebSocket.cpp
@@ -940,7 +940,7 @@ IPAddress AsyncWebSocketClient::remoteIP()
 {
   if (!_client)
   {
-    return IPAddress(0U);
+    return IPAddress((uint32_t)0);
   }
   return _client->remoteIP();
 }
